update(Install): Clarification on Halyard Install
The step I believe we had before with regards to the permissions on the directories seems to have been removed when the documentation was last updated...  But we see it happen almost all the time with all junior candidates we hire, that they get stuck on these permissions.

Added this information so that people don't hit an error (which is very good at explaining what needs to be done) and then make the change after.