docs(chore): add release notes templates (#310)
Add partials and layouts for halyard and operator. Both sort by date,
which requires "date: yyyy-mm-dd" in the front matter. The issue with
the list-recent-release-notes is that it sorts pages by weight and then
title, which causes the Halyard release notes to not be in the correct
order because the title is a string. For example, v1.10.0 appears as
older than 1.9.4.