fix(rn): fix front matter in release notes
The leading zero likely had the page being interpreted as a hex number
instead of an int, so the orderering from the pages got out of whack
(2.30 was listed after 2.28). With the logic on the release notes page,
it meant that "28" was considered the latest active patch release and
therefore kicked the "30" version into the deprecated section. Also
fixed a logical bug in the value setting for this page when we release
version 28.0.0 of the product xP